Convert test output to libxo.
All test output is now on stdout.

The SUMMARY output remains on stderr and is not part of the XML or JSON
output as it's easy to produce by parsing the test results.

The contents of the test stdout are only included in reason strings in
verbose mode (-v) as they are better presented in as seperate items in
XML and JSON output.